[norm] #26: five function per file: read_input/event_callbacks/arrows*.c
